Remove failing test on R-devel
You shouldn't expect the size of a function to be the same between lobstr and base R because base R counts the size of each symbol multiple times. This presumably only failed on R 3.6 because the generated byte code is now contains a duplicated symbol.

I've added a bunch of helper functions that will hopefully make debugging this sort of problem easier in the future.